Fixes multiple bugs in linked list and hash map.
Linked list insertion did not work as intended. Links were not updated so
newly inserted elements were not reachable. In addition, the sorted property was not maintained if the inserted element was smaller than the head of the list. This has been fixed, but required changes in the behaviour of vul_list_find as well as vul_list_insert, vul_map_get and vul_map_remove.