SVD works
SVD on dense matrices (square and rectangular) works, as
well as reconstruction based on the resulting basis vectors.

The current algorithm is *slow*, but *simple*:
Repeatedly QR-decompose the transpose of the matrix and build
the U and V matrices by multiplying the resulting Q at each step
until all that is left is the S-matrix.

QR decomposition is currenty performened using householder
transformations (and should keep doing that for dense matrices),
but we also have functions that perform givens rotations or
gram-schmidt decomposition (and they work!) that can be dropped in.
We probably want givens rotations for sparse matrices.