Added robin hood hash map implementation
The old linked-list implementation has a different usage
style (an atrocious one), and is left in the repo for
projects that use it. If starting a new project, use the
vul_hash_map.h though, not vul_hash_map_linked_list.h;
it's both faster, smaller, and more pleasant to use.